Don't default PackageVersion for NuGet Pack task generated packages (#49053)
Specifying a default PackageVersion causes Arcade's calculated
Version not be used and instead the VersionSuffix-less Version be used
which indicates a stable packages. That later fails during publishing to
a non stable feed.

This is a mitigation for the official build breaks and a better
implementation that supports not reving the patch major during servicing
needs to be figured out.