[metadata_update] Fix bounds check error (#49328)
The issue is that the STANDALONESIG bounds check was using a 0-based index, and
mono_metadata_bounds_check_slow was compensating by adding 1.

But that made another call to the bounds check fail: in
mono_class_from_typeref_checked we passed a 1-based index.  So in the case
where a TypeRef was using the last AssemblyRef in a delta, the bound check
would fail.

Fixes https://github.com/dotnet/runtime/issues/49227

Co-authored-by: Ryan Lucia <ryan@luciaonline.net>

Co-authored-by: Ryan Lucia <ryan@luciaonline.net>