Improve liveness for 'STORE_BLK(lcl_var)'. (#48797)
* improve liveness for STORE_BLK

* Improve code for `RehomeArgForFastTailCall`.

* Delete an unnecessary change

While reviewing this change again I found that `fgMarkUseDef` should be called when we visit the local itself, the diffs that I saw from this change were because the LCL_VAR was not marked as `GTF_VAR_DEF`. Now the lclVar is marked as `GTF_VAR_DEF` so we don't need this extra logic.