[wasm] Use response files for `emcc` invocations (#51530)
* [wasm] Use response files for `emcc` invocations

This adds response file for the linker command line, which will be the
longest one. And it uses a known name for the response file.

In future, once we have dependency checking, we can use such response
files for that too.

Fixes #51437 .

* [wasm] Emit emcc arguments, one per line in the rsp file

Having the arguments one per line, also means that we don't have to
quote the file paths. And it is more readable.

* [wasm] Revert to quote file paths even in response file for emcc

Suggested by Pranavkm:
```
.. their format does require whitespaces to be quoted:
https://github.com/emscripten-core/emscripten/blob/main/tools/response_file.py#L16-L53.
```