Unix: Additional check if the RSA key is external.
Currently we are checking the RSA_METHOD to see if an RSA key "external",
where the private key material may not be exportable. This may not work
because the RSA key itself can declare if it is external or not.

This also checks the flags of the RSA key object itself so attempting
to export the private key is not performed.